hlll: support for high precision

+++ hlll.py
@@ -23,12 +23,14 @@ def hlll_wrapper(A, threads=4, threshold=40, **kwds):
     if not os.path.exists("hlll"):
         raise RuntimeError("cannot find `hlll`")
 
+    int_type = 'mpz'
     if isinstance(A, list):
         return_list = True
         m, n = len(A), len(A[0])
     elif isinstance(A, IntegerMatrix):
         return_list = False
         m, n = A.nrows, A.ncols
+        int_type = A.int_type
     else:
         raise TypeError("Matrix `A` type '%s' unknown." % type(A))
     tmpfile = f"hlll_tmp_matrix-{m}-{n}.mat"
@@ -39,15 +41,21 @@ def hlll_wrapper(A, threads=4, threshold=40, **kwds):
     command += f"--delta {kwds.pop('delta', 0.999)} "
     if kwds.pop('verbose', False):
         command += "--verbose "
+    prec = kwds.get("precision")
+    if prec is not None:
+        command += f"--precision {prec} "
+
+    keeptmp = kwds.pop("keeptmp", False)
 
     with open(tmpfile, "w") as f:
         f.write(f"[{str_mat(A)}]")
 
     try:
         subprocess.check_call(command, shell=True)
-        A = IntegerMatrix.from_file(tmpfile, int_type=A.int_type)
+        A = IntegerMatrix.from_file(tmpfile, int_type=int_type)
     finally:
-        os.remove(tmpfile)
+        if not keeptmp:
+            os.remove(tmpfile)
 
     if return_list:
         A = [[A[i, j] for j in range(A.ncols)] for i in range(A.nrows)]
